Refactor logging and update imports
- Renamed the logger module from `logger.py` to `logging.py` for clarity and consistency.
- Updated import statements in `__init__.py` to reflect the new logger module name.
- Introduced a new logging setup with color support for better log visibility.

These changes improve the organization of the logging functionality within the DataViewer library.
