+++ dataviewer/__init__.py
@@ -3,7 +3,7 @@ DataViewer - 一个用于创建嵌套可视化的Python库
 """
 from .components import *
 from .core import ComponentRegistry, Page
-from .logger import logger
+from .logging import logger
 from .renderers import CellImageRenderer, CellRendererRegistry
 
 __all__ = [
